docs: Recyclerview 모아서 하나의 object에 정리
